Style calculator buttons
This took a while because I encountered an error. I was attempting to get my button text to grow in size when the button was hovered over, for this I used the 'transform: scale(2)' property. What I failed to realise however was that this increased the container size of the button text, causing it to overflow to take the space of the other buttons. Hence once I hovered over a button and attempted to hover over a different button, my cursor would have to be far enough away from the first button for the effect to take place. So instead of using transform, I simply increased the font size once the button was hovered, achieving the intended effect without increasing the size of the container.